01_baseline_torch_compile/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
huggingface_hub>=0.24.0
Pillow>=10.0.0
safetensors>=0.4.0
